% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{build_XtX}
\alias{build_XtX}
\title{Compute X'X}
\usage{
build_XtX(X)
}
\arguments{
\item{X}{\code{T x p} matrix of predictors}
}
\value{
Block diagonal \code{Tp x Tp} Matrix (object) where each \code{p x p} block is \code{tcrossprod(matrix(X[t,]))}
}
\description{
Build the \code{Tp x Tp} matrix XtX using the Matrix() package
}
\note{
X'X is a one-time computing cost. Special cases may have more efficient computing options,
but the Matrix representation is important for efficient computations within the sampler.
}
\keyword{internal}
