% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASV.R
\name{computeDIC_ASV}
\alias{computeDIC_ASV}
\title{Function for calculating DIC and Pb (Bayesian measures of model complexity and fit by Spiegelhalter et al. 2002)}
\usage{
computeDIC_ASV(y, beta, post_sigma2, post_loglike)
}
\arguments{
\item{y}{the \code{T x 1} vector of time series observations.}

\item{beta}{the known mean of the process. 0 by default.}

\item{post_sigma2}{posterior samples of the variance, i.e. exp(h)}

\item{post_loglike}{log likelihood based on the posterior sample.}
}
\value{
a list containing DIC and p_d. Two options for estimating both DIC and p_d, which are both included.
}
\description{
Function for calculating DIC and Pb (Bayesian measures of model complexity and fit by Spiegelhalter et al. 2002)
}
\keyword{internal}
