% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ergMean}
\alias{ergMean}
\title{Compute the ergodic (running) mean.}
\usage{
ergMean(x)
}
\arguments{
\item{x}{vector for which to compute the running mean}
}
\value{
A vector \code{y} with each element defined by \code{y[i] = mean(x[1:i])}
}
\description{
Compute the ergodic (running) mean.
}
\keyword{internal}
