% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASV.R
\name{generate_ly2hat}
\alias{generate_ly2hat}
\title{Posterior predictive sampler on the transformed y (log(y^2))}
\usage{
generate_ly2hat(h, p_error_term)
}
\arguments{
\item{h}{the log varaince term h}

\item{p_error_term}{2 dimensional data frame containing mean and the variance from the 10 componenet
Gaussian mixture in Omori et al 2007 paper.}
}
\value{
a vector containing posterior predictive on log(y^2)
}
\description{
Posterior predictive sampler on the transformed y (log(y^2))
}
\keyword{internal}
