% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getEffSize}
\alias{getEffSize}
\title{Summarize of effective sample size}
\usage{
getEffSize(postX)
}
\arguments{
\item{postX}{An array of arbitrary dimension \code{(nsims x ... x ...)}, where \code{nsims} is the number of posterior samples}
}
\value{
Table of summary statistics using the function \code{summary()}.
}
\description{
Compute the summary statistics for the effective sample size (ESS) across
posterior samples for possibly many variables
}
\keyword{internal}
