% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{initChol_spam}
\alias{initChol_spam}
\title{Compute initial Cholesky decomposition for Bayesian Trend Filtering}
\usage{
initChol_spam(nT, D = 1)
}
\arguments{
\item{nT}{number of time points}

\item{D}{degree of differencing (D = 1 or D = 2)}
}
\description{
Computes the Cholesky decomposition for the quadratic term in the (Gaussian) posterior
of the Bayesian Trend Filtering coefficients. The sparsity pattern will not change during the
MCMC, so we can save computation time by computing this up front.
}
\keyword{internal}
