% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{initCholReg.spam}
\alias{initCholReg.spam}
\title{Compute initial Cholesky decomposition for TVP Regression}
\usage{
initCholReg.spam(obs_sigma_t2, evol_sigma_t2, XtX, D = 1)
}
\description{
Computes the Cholesky decomposition for the quadratic term in the (Gaussian) posterior
of the TVP regression coefficients. The sparsity pattern will not change during the
MCMC, so we can save computation time by computing this up front.
}
