% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{initDHS}
\alias{initDHS}
\title{Initialize the evolution error variance parameters}
\usage{
initDHS(omega)
}
\arguments{
\item{omega}{\code{T x p} matrix of evolution errors}
}
\value{
List of relevant components: the \code{T x p} evolution error SD \code{sigma_wt},
the \code{T x p} log-volatility \code{ht}, the \code{p x 1} log-vol unconditional mean(s) \code{dhs_mean},
the \code{p x 1} log-vol AR(1) coefficient(s) \code{dhs_phi},
the \code{T x p} log-vol innovation SD \code{sigma_eta_t} from the PG priors,
the \code{p x 1} initial log-vol SD \code{sigma_eta_0},
and the mean of log-vol means \code{dhs_mean0} (relevant when \code{p > 1})
}
\description{
Compute initial values for evolution error variance parameters under the dynamic horseshoe prior
}
\keyword{internal}
