% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp-package.R
\name{is.dsp}
\alias{is.dsp}
\alias{dsp-class}
\title{Dynamic Shrinkage Process (dsp) Object}
\usage{
is.dsp(object)
}
\arguments{
\item{object}{any \R object}
}
\value{
Logical, whether \code{object} inherits from class \code{dsp}.
}
\description{
Check whether an object is a fitted \code{dsp} model.
}
\details{
A \code{dsp} object is a classified list created by \code{\link{dsp_fit}}.
It contains \code{mcmc_output}, a list of posterior draws whose matrix
elements are \code{nsave} by \code{T} and whose vector elements are of
length \code{nsave}; \code{DIC}, the deviance information criterion and
effective number of parameters; \code{mcpar}, the number of saved draws,
the burn-in and the thinning interval; and \code{model_spec}, the
\code{dsp_spec} object describing the fitted model.
}
\seealso{
\code{\link[=as.mcmc.dsp]{as.mcmc.dsp()}}, \code{\link[=summary.dsp]{summary.dsp()}}
}
