% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{logit}
\alias{logit}
\title{Compute the log-odds}
\usage{
logit(x)
}
\arguments{
\item{x}{scalar or vector in (0,1) for which to compute the (componentwise) log-odds}
}
\value{
A scalar or vector of log-odds
}
\description{
Compute the log-odds
}
\keyword{internal}
