% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make.signal}
\alias{make.signal}
\title{Create the functions}
\usage{
make.signal(name, n = 1024, snr = Inf)
}
\arguments{
\item{name}{string matching the "name" argument in the \code{make.signal()} function,
e.g. "bumps" or "doppler" (see below)}

\item{n}{number of points}

\item{snr}{signal-to-noise ratio; default is \code{Inf}}
}
\value{
vector of simulated y-values
}
\description{
Replace the \code{make.signal()} call in the \code{wmtsa}
package (no longer available)
}
\details{
The function names include "dirac", "kronecker", "heavisine", "bumps", "blocks",
"doppler", "ramp", "cusp", "crease", "sing", "hisine",
"losine", "linchirp", "twochirp", "quadchirp",
"mishmash1", "mishmash2", "mishmash3", "levelshift",
"jumpsine", "gauss", "linear", "quadratic", "cubic"
}
