% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_cp}
\alias{plot_cp}
\title{Plot the series with change points}
\usage{
plot_cp(mu, cp_inds)
}
\arguments{
\item{mu}{the \code{T x 1} vector of time series observations (or fitted values)}

\item{cp_inds}{the \code{n_cp x 1} vector of indices at which a changepoint is identified}
}
\description{
Plot the time series with distinct segments identified by color.
}
