% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_fitted}
\alias{plot_fitted}
\title{Plot the Bayesian trend filtering fitted values}
\usage{
plot_fitted(y, mu, postY, y_true = NULL, t01 = NULL)
}
\arguments{
\item{y}{the \code{T x 1} vector of time series observations}

\item{mu}{the \code{T x 1} vector of fitted values, i.e., posterior expectation of the state variables}

\item{postY}{the \code{nsims x T} matrix of posterior draws from which to compute intervals}

\item{y_true}{the \code{T x 1} vector of points along the true curve}

\item{t01}{the observation points; if NULL, assume \code{T} equally spaced points from 0 to 1}
}
\description{
Plot the BTF posterior means with posterior credible intervals (pointwise and joint),
the observed data, and true curves (if known)
}
\examples{
simdata = simUnivariate(signalName = "doppler", T = 128, RSNR = 7, include_plot = FALSE)
y = simdata$y
out = btf(y)
plot_fitted(y, mu = colMeans(out$mu), postY = out$yhat, y_true = simdata$y_true)
}
