% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{post_spec_dsp}
\alias{post_spec_dsp}
\title{Compute the posterior distribution of the spectrum of a TVAR(p) model}
\usage{
post_spec_dsp(post_ar_coefs, post_sigma_e, n.freq = 500)
}
\arguments{
\item{post_ar_coefs}{(nsave x T x p) array of TVAR(p) coefficients}

\item{post_sigma_e}{(nsave x 1) vector of observation standard deviation}

\item{n.freq}{number of frequencies at which to evaluate the spectrum}
}
\value{
A list containing (1) the vector of frequencies at which the spectrum
was evaluated and (2) a (nsave x T x n.freq) array of the spectrum values
for each MCMC simulation at each time.
}
\description{
Compute the posterior distribution of the spectrum of a TVAR(p) model
}
