% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{print.dsp}
\alias{print.dsp}
\title{Print a summary of dsp object}
\usage{
print.dsp(object, ...)
}
\arguments{
\item{object}{object of class dsp from \code{\link[=dsp_fit]{dsp_fit()}}}

\item{...}{currently unused}
}
\description{
Print a summary of dsp object
}
\details{
A brief summary of the settings used to fit the model including number of iterations,
burn in, and thinning rates.
}
\examples{

signal = c(rep(0, 50), rep(10, 50))
noise = rep(1, 100)
noise_var = rep(1, 100)
for (k in 2:100){
  noise_var[k] = exp(0.9*log(noise_var[k-1]) + rnorm(1, 0, 0.5))
  noise[k] = rnorm(1, 0, sqrt(noise_var[k])) }

y = signal + noise
model_spec = dsp_spec(family = "gaussian", model = "changepoint")
mcmc_output = dsp_fit(y, model_spec = model_spec)
print(mcmc_output)



}
