% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rig}
\alias{rig}
\title{Sample from an inverse-Gaussian distribution}
\usage{
rig(n, mean, scale)
}
\arguments{
\item{n}{the number of deviates required. If this has length > 1 then the length is taken as the number of deviates required.}

\item{mean}{vector of mean values.}

\item{scale}{vector of scale parameter values (lambda)}
}
\description{
Using code from the \code{mgcv} package
}
