% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASV.R
\name{sample_jfast}
\alias{sample_jfast}
\title{Posterior Sample from 10-componenets Mixture Gaussian based on Omori et al. 2007}
\usage{
sample_jfast(T, obs = NULL)
}
\arguments{
\item{T}{the number of observations}

\item{obs}{the observed value, should be T-length long}
}
\value{
\code{T x 2} data frame containing the posterior mean on the "mean" column
and the posterior variance on the "variance" column based on the provided obs.
}
\description{
Posterior Sample from 10-componenets Mixture Gaussian based on Omori et al. 2007
}
\note{
If the "obs" are not provided componenets are sampled from the prior distribution.
}
