% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_samplers.R
\name{sampleBTF}
\alias{sampleBTF}
\title{Sampler for first or second order random walk (RW) Gaussian dynamic linear model (DLM)}
\usage{
sampleBTF(
  y,
  obs_sigma_t2,
  evol_sigma_t2,
  D = 1,
  loc_obs = NULL,
  chol0 = NULL,
  prior_mean = NULL
)
}
\arguments{
\item{y}{the \code{T x 1} vector of time series observations}

\item{obs_sigma_t2}{the \code{T x 1} vector of observation error variances}

\item{evol_sigma_t2}{the \code{T x 1} vector of evolution error variances}

\item{D}{the degree of differencing (one or two)}

\item{loc_obs}{list of the row and column indices to fill in a band-sparse matrix}

\item{chol0}{(optional) the \code{m x m} matrix of initial Cholesky factorization;
if NULL, use the \code{Matrix} package for sampling, otherwise use the \code{spam} package}

\item{prior_mean}{optional (default is NULL); numeric \code{T x 1} vector specifying the prior mean of \code{mu}}
}
\value{
\code{T x 1} vector of simulated states
}
\description{
Compute one draw of the \code{T x 1} state variable \code{mu} in a DLM using back-band substitution methods.
This model is equivalent to the Bayesian trend filtering (BTF) model, assuming appropriate
(shrinkage/sparsity) priors for the evolution errors.
}
\note{
Missing entries (NAs) are not permitted in \code{y}. Imputation schemes are available.
}
\keyword{internal}
