% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{simUnivariate}
\alias{simUnivariate}
\title{Simulate noisy observations from a function}
\usage{
simUnivariate(signalName = "bumps", T = 200, RSNR = 10,
  include_plot = TRUE)
}
\arguments{
\item{signalName}{string matching the "name" argument in the \code{make.signal()} function,
e.g. "bumps" or "doppler"}

\item{T}{number of points}

\item{RSNR}{root-signal-to-noise ratio}

\item{include_plot}{logical; if TRUE, include a plot of the simulated data and the true curve}
}
\value{
a list containing
\itemize{
\item the simulated function \code{y}
\item the true function \code{y_true}
\item the true observation standard devation \code{sigma_true}
}
}
\description{
Builds upon the \code{make.signal()} function in the \code{wmtsa} package
to include Gaussian noise with a user-specificied root-signal-to-noise ratio.
}
\note{
The root-signal-to-noise ratio is defined as RSNR = [sd of true function]/[sd of noise].
}
\examples{
sims = simUnivariate() # default simulations
names(sims) # variables included in the list

}
