% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{summary.dsp}
\alias{summary.dsp}
\title{Summarize DSP MCMC chains}
\usage{
summary.dsp(object, pars, probs = c(0.025, 0.25, 0.5, 0.75, 0.975), ...)
}
\arguments{
\item{object}{object of class dsp from \code{\link[=dsp_fit]{dsp_fit()}}}

\item{pars}{parameter names specified for summaries; currently defaults to all parameters named in object$mcmc_output}

\item{probs}{numeric vector of \code{\link[=quantile]{quantile()}}s requested for posterior summary of pars. Defaults to c(0.025, 0.25, 0.50, 0.75, 0.975)}

\item{...}{currently unused}
}
\value{
Currently, returns a named list of the same length as pars where within each element of the list
is a numeric matrix of vector. For matrices, each row is a time point of the parameter and each column
is a named summary. For vectors (scalar parameters), each element is a named summary.
}
\description{
Summarize DSP MCMC chains
}
