% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abco.R
\name{t_sampleSVparams}
\alias{t_sampleSVparams}
\title{Sampler for the stochastic volatility parameters}
\usage{
t_sampleSVparams(omega, svParams)
}
\arguments{
\item{omega}{\code{T} vector of errors}

\item{svParams}{list of parameters to be updated}
}
\value{
List of relevant components in \code{svParams}: \code{sigma_wt}, the \code{T} vector of standard deviations,
and additional parameters associated with SV model.
}
\description{
Compute one draw of the normal stochastic volatility parameters.
The model assumes an AR(1) for the log-volatility.
}
\keyword{internal}
