% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp-package.R
\docType{package}
\name{dsp-package}
\alias{dsp}
\alias{dsp-package}
\title{dsp: Dynamic Shrinkage Process and Change Point Detection}
\description{
A full, efficient MCMC sampling algorithm for dynamic shrinkage processes (DSPs). DSPs extend popular global-local shrinkage priors, such as the horseshoe prior for sparse signals, to the time series setting by allowing the shrinkage behavior to depend on the history of the shrinkage process. The resulting processes are locally adaptive, which is important for time series data and regression functions with irregular features. The package provides the component samplers for the Gibbs sampler for DSPs, as well as a full MCMC implementation for Bayesian trend filtering (BTF) with dynamic horseshoe processes as the prior (penalty). BTF estimates are used for curve-fitting of univariate data, typically with irregular features. The BTF model is implemented using a dynamic linear model (DLM) framework, which provides efficient computations and a platform for useful extensions. BTF penalizes differences (first or second, in this case) of the conditional expectation (i.e., the signal) to produce approximately locally constant or locally linear estimates. We use DSPs as the prior on the differences, which produces curve estimates and credible bands that are highly adaptive to both rapidly- and slowly-changing features. We also provide BTF model implementations for the (static) horseshoe (HS) prior and a normal- inverse-Gamma (NIG) prior. In all cases, computations are linear in the number of time points, so the MCMC samplers are highly efficient. Besides curve-fitting via BTF, we include full, efficient MCMC sampling algorithms for dynamic shrinkage processes applied to (1) dynamic regression with time-varying coefficients and (2) B-spline models for curve-fitting. In the dynamic regression model, we regress a dynamic (scalar) response on a vector of dynamic predictors for which the corresponding regression coefficients are time-varying. The first or second differences of the regression coefficients are penalized using DSPs (with options for HS and NIG priors), allowing for highly adaptive regression functions. In the B-spline setting, we penalize the differences of the B-spline basis coefficients, similar to P-splines, using DSPs (with options for HS and NIG priors). The resulting curve-fitting model is highly adaptive, like the BTF model above, but easily incorporates unequally-spaced observation points. Change points are estimated using a threshold parameter. The model can deal with heteroscedastic measurement error, and jointly identifies both local outliers and change points. Methods in this pacakge are described in Kowal et al. (2019) \doi{10.1111/rssb.12325}, Wu et al. (2024) \doi{10.1080/07350015.2024.2362269}, Schafer and Matteson (2024) \doi{10.1080/00401706.2024.2407316}, and Cho and Matteson (2024) \href{https://arxiv.org/abs/2408.11315}{arXiv:2408.11315}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/schafert/dsp}
  \item Report bugs at \url{https://github.com/schafert/dsp/issues}
}

}
\author{
\strong{Maintainer}: Toryn Schafer \email{toryn27@gmail.com} (\href{https://orcid.org/0000-0001-5594-7697}{ORCID})

Authors:
\itemize{
  \item Daniel R. Kowal \email{daniel.kowal@rice.edu} [copyright holder]
  \item Haoxuan Wu \email{hw399@cornell.edu}
  \item Jason B. Cho \email{bc454@cornell.edu}
  \item David S. Matteson \email{matteson@cornell.edu}
}

}
\keyword{internal}
