% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_samplers.R
\name{dsp_fit}
\alias{dsp_fit}
\title{MCMC Sampler for Dynamic Shrinkage with Changepoint}
\usage{
dsp_fit(
  y,
  family = "gaussian",
  trend = NULL,
  cp = FALSE,
  evol_error = "DHS",
  D = 1,
  obsSV = "const",
  useAnom = FALSE,
  nsave = 1000,
  nburn = 1000,
  nskip = 4,
  mcmc_params = list("mu", "omega", "r"),
  computeDIC = TRUE,
  verbose = FALSE,
  cp_thres = 0.4,
  ...
)
}
\arguments{
\item{y}{a numeric vector of the \code{T x 1} vector of time series observations}

\item{family}{a string specifying exponential family for observation likelihood.
Defaults to "gaussian" and implementation also available for "negbinomial"}

\item{trend}{optional argument specifying form of the trend for family = "gaussian"; default NULL indicates trend is a dynamic conditional mean (i.e., Bayesian trend filtering) otherwise
\itemize{
\item "zero_error" ... for sparse trend
\item a named list elements "times": the \code{T x 1} vector of observation points;
if NULL, assume equally spaced and
"num_knots" the number of knots; if NULL, use the default
of \code{max(20, min(ceiling(T/4), 150))} with bspline trend;
\item a \code{T x p} matrix of time series predictors for a regression trend;
}}

\item{cp}{a logical flag (default is FALSE) indicating to determine whether to use threshold shrinkage with changepoints; only implemented for family = "gaussian".}

\item{evol_error}{the evolution error distribution; must be one of
\itemize{
\item the dynamic horseshoe prior ('DHS');
\item the static horseshoe prior ('HS');
\item the Bayesian lasso ('BL');
\item the normal stochastic volatility model ('SV');
\item the normal-inverse-gamma prior ('NIG').
}}

\item{D}{integer scalar indicating degree of differencing defaults to 1; implementation is available D = 0, D = 1, or D = 2}

\item{obsSV}{Options for modeling the error variance for family = "gaussian". It must be one of the following:
\itemize{
\item const: Constant error variance for all time points.
\item SV: Stochastic Volatility model.
\item ASV: Adaptive Stochastic Volatility model.
}}

\item{useAnom}{logical (default FALSE); if TRUE, include an anomaly component in the observation equation
(only for threshold shrinkage with changepoints and family = "gaussian".)}

\item{nsave}{integer scalar (default = 1000); number of MCMC iterations to record}

\item{nburn}{integer scalar (default = 1000); number of MCMC iterations to discard (burn-in)}

\item{nskip}{integer scalar (default = 4); number of MCMC iterations to skip between saving iterations,
i.e., save every (nskip + 1)th draw}

\item{mcmc_params}{list of character scalars naming parameters for which we store the MCMC output;
must be one or more of:
\itemize{
\item "mu" (conditional mean)
\item "yhat" (posterior predictive distribution)
\item "evol_sigma_t2" (evolution error variance)
\item "obs_sigma_t2" (observation error variance)
\item "dhs_phi" (DHS AR(1) coefficient)
\item "dhs_mean" (DHS AR(1) unconditional mean)
\item "r" (overdispersion when family = "negbinomial")
}
defaults to list("mu", "omega", "r")}

\item{computeDIC}{logical; if TRUE (default), compute the deviance information criterion \code{DIC}
and the effective number of parameters \code{p_d}}

\item{verbose}{logical; should extra information on progress be printed to the console? Defaults to FALSE}

\item{cp_thres}{Proportion of posterior samples of latent indicator being 1 needed to declare a changepoint; defaults to 0.4}

\item{...}{optional additional arguments to pass to \code{\link{btf_nb}} when family = "negbinomial"}
}
\value{
\code{dsp_fit} returns an object of class "\code{dsp}".

An object of class "\code{dsp}" is defined as a list containing at least the following components:
\item{pars}{a list of the \code{nsave} MCMC samples for the parameters named in \code{mcmc_params}} ## TODO change so matches
\item{cp}{if threshold shrinkage with changepoints is used, also return detected changepoint locations; otherwise FALSE}
\item{DIC}{}
\item{family}{value supplied for family argument}
\item{evol_error}{value supplied for evol_error argument}
\item{D}{value supplied for D argument}
\item{obsSV}{value supplied for obsSV argument}
\item{mcpar}{named vector of supplied nsave, nburn, and nskip}
\item{cp_thres}{value supplied for cp_thres argument}
}
\description{
Run the MCMC for Bayesian trend filtering with a penalty on zeroth (D = 0),
first (D = 1), or second (D = 2) differences of the conditional expectation.
The user has the option to utilize threshold shrinkage to identify changepoints
for D = 1 or 2.
The penalty is determined by the prior on the evolution errors, which include:
\itemize{
\item the dynamic horseshoe prior ('DHS');
\item the static horseshoe prior ('HS');
\item the Bayesian lasso ('BL');
\item the normal stochastic volatility model ('SV');
\item the normal-inverse-gamma prior ('NIG').
}
In each case, the evolution error is a scale mixture of Gaussians.
Note that only 'DHS' works for threshold shrinkage with changepoints.
Sampling is accomplished with a (parameter-expanded) Gibbs sampler,
mostly relying on a dynamic linear model representation.
}
\note{
The data \code{y} may contain NAs, which will be treated with a simple imputation scheme
via an additional Gibbs sampling step. In general, rescaling \code{y} to have unit standard
deviation is recommended to avoid numerical issues when family is "gaussian".
}
\examples{

# Example 1: Change in mean with stochastic volatility
signal = c(rep(0, 50), rep(10, 50))
noise = rep(1, 100)
noise_var = rep(1, 100)
for (k in 2:100){
  noise_var[k] = exp(0.9*log(noise_var[k-1]) + rnorm(1, 0, 0.5))
  noise[k] = rnorm(1, 0, sqrt(noise_var[k])) }

y = signal + noise
mcmc_output = dsp_fit(y, cp=TRUE, mcmc_params = list('yhat', 'mu', "omega", "r"))
cp = mcmc_output$cp
print(paste0('Changepoint Locations: ', cp))
plot(mcmc_output, y, y_true = signal)

# Example 2: Change in mean with Outliers
signal = c(rep(0, 100), rep(5, 100))
y = c(rep(0, 100), rep(5, 100)) + rnorm(200)
y[50] = 10
y[150] = -10

mcmc_output = dsp_fit(y, cp=TRUE, useAnom = TRUE, mcmc_params = list('yhat', 'mu', "omega", "r"))
cp = mcmc_output$cp
plot(mcmc_output, y, y_true = signal)

# Example 3: Change in linear trend
signal = c(seq(1, 50), seq(51, 2))
y = c(seq(1, 50), seq(51, 2)) + rnorm(100)

mcmc_output = dsp_fit(y, cp=TRUE, D=2, mcmc_params = list('yhat', 'mu', "omega", "r"))
cp = mcmc_output$cp
plot(mcmc_output, y,y_true = signal)


}
