% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp_model.R
\name{dsp_spec}
\alias{dsp_spec}
\title{Model Specification}
\usage{
dsp_spec(family, model, ...)
}
\arguments{
\item{family}{A character string specifying the model family. Must be one of:
\itemize{
\item "gaussian": Gaussian family.
\item "negbinom": Negative binomial family.
}}

\item{model}{A character string specifying the model type:
\itemize{
\item \code{family} = "gaussian":
\itemize{
\item "changepoint": Change point detection with
Adaptive Bayesian Changepoint analysis and local Outlier (ABCO),
\item "smoothing": Bayesian smoothing,
\item "regression": Time-varying regression,
\item "bspline": Bayesian smoothing with B-spline for irregularly spaced or functional time-series.
}
\item \code{family} = "negbinom":
\itemize{
\item "smoothing": Bayesian smoothing.
}
}}

\item{...}{Additional arguments based on \code{family} and \code{model}.
For \code{family} = "gaussian" and \code{model} = "regression", \code{X} is required.
Other arguments are optional and default to predefined value.

\tabular{lll}{
\strong{Family}    \tab  \strong{Model}       \tab    \strong{Required Arguments}  \cr
"gaussian"    \tab "changepoint"     \tab \code{D}, \code{useAnom}, \code{obsSV} \cr
"gaussian"    \tab "smoothing"       \tab \code{D}, \code{evol_error}, \code{obsSV}, \code{zero_error}, \code{D_asv}, \code{evol_error_asv}, \code{nugget_asv}\cr
"gaussian"    \tab "regression"      \tab \code{D}, \code{evol_error}, \code{obsSV}, \code{X}, \code{D_asv}, \code{evol_error_asv}, \code{nugget_asv} \cr
"gaussian"    \tab "bspline"        \tab \code{D}, \code{evol_error}, \code{obsSV}, \code{times}, \code{num_knots}  \cr
"negbinom"    \tab "smoothing"       \tab \code{D}, \code{evol_error}, \code{r_init}, \code{r_sample}, \code{offset}  \cr
}

\itemize{
\item \strong{Shared Arguments}:
\itemize{
\item \code{D}: integer scalar indicating degree of differencing.
Implementation is available D = 0, 1, or 2 (default) for \code{family} = "gaussian",
and D = 1 and 2 (default) for \code{family} = "negbinomial".
\item \code{obsSV}:Options for modeling the error variance for \code{family} = "gaussian".
It must be one of the following:
\itemize{
\item "const": Constant error variance for all time points (default);
\item "SV": Stochastic Volatility model;
\item "ASV": Adaptive Stochastic Volatility model;
}
\item \code{D_asv}: integer; degree of differencing (0, 1 (default), or 2) for the ASV model. Only used when \code{obsSV = "ASV"}.
\item \code{evol_error_asv}: character; "HS" by default. evolution error distribution for the ASV model. Must be one of the five options used in \code{evol_error}. Only used when \code{obsSV = "ASV"}
\item \code{nugget_asv}: logical; if \code{TRUE} (default), fits the nugget variant of the ASV model. Only used when \code{obsSV = "ASV"}.
\item \code{evol_error}: the evolution error distribution; .
\itemize{
\item the dynamic horseshoe prior ("DHS") (default);
\item the static horseshoe prior ("HS");
\item the Bayesian lasso ("BL") (\code{family} = "gaussian" only);
\item the normal stochastic volatility model ("SV") (\code{family} = "gaussian" only);
\item the normal-inverse-gamma prior ("NIG") (\code{family} = "gaussian" only).
}
}
\item \strong{Model-specific Arguments}:
\itemize{
\item \code{model} = "changepoint":
\itemize{
\item \code{useAnom}: logical; Defaults to FALSE. if TRUE, include an anomaly component
in the observation equation.
}
\item \code{family} = "gaussian", \code{model} = "smoothing":
\itemize{
\item \code{zero_error}: one of the 5 distributions listed in \code{"evol_error"};
Additional penalty distribution for the conditional expectation i.e., shrinkage to zero (optional).
}
\item \code{family} = "gaussian", \code{model} = "regression":
\itemize{
\item \code{X}: matrix or data frame; T x p matrix of time series predictors
}
\item \code{model} = "bspline":
\itemize{
\item \code{times}: length \code{T} vector; observation indices; if NULL (Default), assume equally spaced.
\item \code{num_knots}: numeric; the number of knots to be used for bspline. defaults to 20.
}
\item \code{family} = "negbinomial", \code{model} = "smoothing":
\itemize{
\item \code{r_init}: numeric; initial value (defaults to 5) of MCMC sampling for
overdispersion parameter.
\item \code{r_sample}: logical; If \code{TRUE}, the overdispersion is sampled,
if \code{FALSE} (default), fixed at "\code{r_init}.
\item \code{offset}: length \code{T} vector; offset values for the log conditional expectation
}
}
}}
}
\value{
A list containing the model specification.
}
\description{
Method for creating dsp specification object prior to fitting.
}
\examples{
model_spec <- dsp_spec(family = "gaussian",
                      model = "changepoint")


}
