% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getARpXmat}
\alias{getARpXmat}
\title{Compute the design matrix X for AR(p) model}
\usage{
getARpXmat(y, p = 1, include_intercept = FALSE)
}
\arguments{
\item{y}{(T x 1) vector of responses}

\item{p}{order of AR(p) model}

\item{include_intercept}{logical; if TRUE, first column of X is ones}
}
\description{
Compute the design matrix X for AR(p) model
}
