% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{initEvol0}
\alias{initEvol0}
\title{Initialize the parameters for the initial state variance}
\usage{
initEvol0(mu0)
}
\arguments{
\item{mu0}{\code{p x 1} vector of initial values (undifferenced)}
}
\value{
List of relevant components: the \code{p x 1} evolution error SD \code{sigma_w0},
the \code{p x 1} precisions \code{tau_j0}, and the \code{p x 1} parameter-expanded RV's \code{xi_j0}
}
\description{
The initial state SDs are assumed to follow standard half-Cauchy priors, C+(0,1).
This function initalizes the parameters for a PX-Gibbs sampler.
}
