% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{invlogit}
\alias{invlogit}
\title{Compute the inverse log-odds}
\usage{
invlogit(x)
}
\arguments{
\item{x}{scalar or vector for which to compute the (componentwise) inverse log-odds}
}
\value{
A scalar or vector of values in (0,1)
}
\description{
Compute the inverse log-odds
}
\examples{
x = seq(-5, 5, length.out = 10^3)
plot(x, invlogit(x))
}
