% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp-package.R
\name{is.dsp}
\alias{is.dsp}
\alias{dsp-class}
\title{Dynamic Shrinkage Process (dsp) Object}
\usage{
is.dsp(object)
}
\arguments{
\item{object}{any \R object}
}
\value{
Logical, whether \code{object} inherits from class \code{dsp}.
}
\description{
Check whether an object is a fitted \code{dsp} model.
}
\details{
A \code{dsp} object is a clcode{\link{dsp_fit}}.
It contains \code{mcmc_output}, a list of posterior draws whose matrix
elements are \code{nsave} br elements are of
length \code{nsave}; \code{DIC}, the deviance information criterion and
effective number of parametber of saved draws,
the burn-in and the thinning interval; and \code{model_spec}, the
\code{dsp_spec} object desc
}
