% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot.dsp}
\alias{plot.dsp}
\title{Plot the Bayesian trend filtering fitted values}
\usage{
\method{plot}{dsp}(mcmc_output, y, y_true = NULL, t01 = NULL, include_joint_bands = FALSE)
}
\arguments{
\item{mcmc_output}{an object of class "acf" with parameter names 'mu' and 'yhat'}

\item{y}{the \code{T x 1} vector of time series observations}

\item{t01}{the observation points; if NULL, assume \code{T} equally spaced points from 0 to 1}

\item{include_joint_bands}{logical; if TRUE, compute simultaneous credible bands}
}
\description{
Plot the BTF posterior means with posterior credible intervals (pointwise and joint),
the observed data, and true curves (if known)
}
\examples{
\dontrun{
simdata = simUnivariate(signalName = "doppler", nT = 128, RSNR = 7, include_plot = FALSE)
y = simdata$y
out = btf(y) # TODO change this so that it uses abco
plot_fitted(y, mcmc_output = out, y_true = simdata$y_true)
}

}
