% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmethods.R
\name{plot.dsp}
\alias{plot.dsp}
\title{Plot the Bayesian trend filtering fitted values}
\usage{
\method{plot}{dsp}(x, y, mu_true = NULL, t01 = NULL, include_joint_bands = FALSE, ...)
}
\arguments{
\item{x}{an object of class 'dsp' from \code{\link[=dsp_fit]{dsp_fit()}} with parameter name 'mu'}

\item{y}{the \code{T x 1} vector of time series observations}

\item{mu_true}{(defaults to NULL) the \code{T x 1} vector of the true conditional mean}

\item{t01}{the observation points; if NULL, assume \code{T} equally spaced points from 0 to 1}

\item{include_joint_bands}{logical; if TRUE, compute simultaneous credible bands}
}
\description{
Plot the BTF posterior mean of the conditional expectation with posterior credible intervals (pointwise and joint),
the observed data, and true curves (if known)
}
\details{
The credible intervals plotted depends on whether \code{x} contains predictions
of the data, i.e., it contains the name 'yhat'. If it does, then a plot is created with
credible intervals for the posterior prediction of y, else credible intervals are calculated
for the conditional expectation contained in the named slot 'mu'.
}
\examples{
signal = c(rep(0, 50), rep(10, 50))
noise = rep(1, 100)
noise_var = rep(1, 100)
for (k in 2:100){
  noise_var[k] = exp(0.9*log(noise_var[k-1]) + rnorm(1, 0, 0.5))
  noise[k] = rnorm(1, 0, sqrt(noise_var[k])) }

y = signal + noise
model_spec = dsp_spec(family = "gaussian", model = "changepoint",
                      mcmc_params = list('yhat', 'mu', "omega", "r"))
mcmc_output = dsp_fit(y, model_spec = model_spec)
plot(mcmc_output, y, mu_true = signal) # plots prediction intervals

}
