% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{print.dsp}
\alias{print.dsp}
\title{Print a summary of dsp object}
\usage{
print.dsp(object, ...)
}
\arguments{
\item{object}{object of class dsp from \code{\link[=dsp_fit]{dsp_fit()}}}

\item{...}{currently unused}
}
\description{
Print a summary of dsp object
}
