% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dsp}
\alias{print.dsp}
\alias{print.dsp_spec}
\title{Print a summary of a model specification or fitted dsp object}
\usage{
\method{print}{dsp}(x, ...)

\method{print}{dsp_spec}(x, ...)
}
\arguments{
\item{x}{object of class dsp_spec from \code{\link[=dsp_spec]{dsp_spec()}}}

\item{...}{currently not used}
}
\description{
Prints basic information about the MCMC sampling settings for the fitted model

Prints basic information about the model specification
}
\details{
A brief summary of the settings used to fit the model including number of iterations,
burn in, and thinning rates.
}
\examples{
set.seed(200)
signal = c(rep(0, 50), rep(10, 50))
noise = rep(1, 100)
noise_var = rep(1, 100)
for (k in 2:100){
  noise_var[k] = exp(0.9*log(noise_var[k-1]) + rnorm(1, 0, 0.5))
  noise[k] = rnorm(1, 0, sqrt(noise_var[k])) }

y = signal + noise
model_spec = dsp_spec(family = "gaussian", model = "changepoint",
                      D = 1, useAnom = TRUE)
mcmc_output = dsp_fit(y, model_spec = model_spec)
print(mcmc_output)

model_spec = dsp_spec(family = "gaussian", model = "changepoint")
print(model_spec)

}
