% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_samplers.R
\name{sample_j_wrap}
\alias{sample_j_wrap}
\title{Sample the conditional posterior distribution on the Gaussian Mixture component
for approximating the log(\chi^2) distribution based on Omori et al. 2007}
\usage{
sample_j_wrap(Td, obs = NULL)
}
\arguments{
\item{Td}{length of the vector}

\item{obs}{\code{Td x 1} vector for the data.}
}
\value{
Dataframe containing the posterior samples: mean and variance for the mixture component.
}
\description{
Sample the conditional posterior distribution on the Gaussian Mixture component
for approximating the log(\chi^2) distribution based on Omori et al. 2007
}
\note{
When the obs is not not specified, the componenets are samples from the prior distribution.
}
