% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_samplers.R
\name{sample_j_wrap}
\alias{sample_j_wrap}
\title{Sampling from 10-component Gaussian Mixture component described in Omori et al. 2007}
\usage{
sample_j_wrap(Td, obs = NULL)
}
\arguments{
\item{Td}{length of the vector}

\item{obs}{\code{Td x 1} vector for the data.}
}
\value{
Dataframe containing the posterior samples: mean and variance for the mixture component.
}
\description{
Samples from the conditional posterior distribution of the \eqn{log(\chi^2)} distribution
by approximating it with the mixture Gaussian distribution described in Omori et al. 2007.
}
\note{
When the obs is not not specified, the components are samples from the prior distribution.
}
