% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_samplers.R
\name{sampleBTF}
\alias{sampleBTF}
\title{Sampler for first or second order random walk (RW) Gaussian dynamic linear model (DLM)}
\usage{
sampleBTF(y, obs_sigma_t2, evol_sigma_t2, D = 1)
}
\arguments{
\item{y}{the \code{T x 1} vector of time series observations}

\item{obs_sigma_t2}{the \code{T x 1} vector of observation error variances}

\item{evol_sigma_t2}{the \code{T x 1} vector of evolution error variances}

\item{D}{the degree of differencing (one or two)}
}
\value{
\code{T x 1} vector of simulated states
}
\description{
Compute one draw of the \code{T x 1} state variable \code{mu} in a DLM using back-band substitution methods.
This model is equivalent to the Bayesian trend filtering (BTF) model, assuming appropriate
(shrinkage/sparsity) priors for the evolution errors.
}
\note{
Missing entries (NAs) are not permitted in \code{y}. Imputation schemes are available.
}
\examples{
# Simulate some data:
T = 1000
y = seq(0, 10, length.out = T) + rnorm(T)
plot(y) # plot the data

obs_sigma_t2 = rep(1, T)  # These could be static or dynamic
evol_sigma_t2 = rep(0.001, T)

# Simulate one draw of the states:
mu = sampleBTF(y = y, obs_sigma_t2, evol_sigma_t2, D = 1)
lines(mu, lwd=8, col='blue') # add the states to the plot

}
