% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{spec_dsp}
\alias{spec_dsp}
\title{Compute the spectrum of an AR(p) model}
\usage{
spec_dsp(ar_coefs, sigma_e, n.freq = 500)
}
\arguments{
\item{ar_coefs}{(p x 1) vector of AR(p) coefficients}

\item{sigma_e}{observation standard deviation}

\item{n.freq}{number of frequencies at which to evaluate the spectrum}
}
\value{
A (n.freq x 2) matrix where the first column is the frequencies
and the second column is the spectrum evaluated at that frequency
}
\description{
Compute the spectrum of an AR(p) model
}
