% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dsp}
\alias{summary.dsp}
\title{Summarize DSP MCMC chains}
\usage{
\method{summary}{dsp}(object, pars, probs = c(0.025, 0.25, 0.5, 0.75, 0.975), ...)
}
\arguments{
\item{object}{object of class dsp from \code{\link[=dsp_fit]{dsp_fit()}}}

\item{pars}{parameter names specified for summaries; currently defaults to all parameters named in object$mcmc_output}

\item{probs}{numeric vector of \code{\link[=quantile]{quantile()}}s requested for posterior summary of pars. Defaults to c(0.025, 0.25, 0.50, 0.75, 0.975)}

\item{...}{currently not being used}
}
\value{
Returns a named list of the same length as pars where within each element of the list
is a numeric matrix (vector parameters) or vector (scalar parameters). For matrices, each row is a time point (or dimension) of the parameter and each column
is a named summary. The names are accessible with \code{colnames}. For vectors (scalar parameters), each element is a named summary.
}
\description{
Summarize DSP MCMC chains
}
\examples{
set.seed(200)
n <- 150
mu <- simUnivariate(name = "quadratic", n = n)        # Underlying trend with a ramp structure
var  <- simUnivariate(name = "bumps", n = n)/2       # Time-varying standard deviation with bumps
y <- rnorm(n = n, mean = mu, sd = sqrt(var))    # Observed data based on above.
# Specify DSP model with ASV observation noise and Horseshoe prior on evolution error
spec <- dsp_spec(family = "gaussian",
                 model = "smoothing",
                 obsSV = "ASV",
                 D_asv = 2)
# Fit the model (Note: longer MCMC runs may be required for stable inference)
fit <- dsp_fit(y, spec, nsave = 500, nburn = 1000)
summary_fit <- summary(fit)
summary_fit$mu[,"mean"]  # contains estimated posterior mean
summary_fit$h[,"mean"] # conrtains log volatility

}
